% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R,
%   R/metaforest-deprecated.r
\name{metaforest-deprecated}
\alias{metaforest-deprecated}
\alias{ClusterMF}
\title{Deprecated functions in package \pkg{metaforest}.}
\usage{
ClusterMF(...)
}
\description{
Metaforest is actively maintained, and I strive to improve its
functionality and ease of use. This sometimes means that a function can, or
must, be deprecated.

The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("-deprecated")}.
}
\section{\code{ClusterMF}}{

For \code{ClusterMF}, use \code{\link{MetaForest}}.
}

\keyword{internal}
