% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDP_interaction.R
\name{PDP_interaction}
\alias{PDP_interaction}
\title{Partial dependence plots split by factor moderator}
\usage{
PDP_interaction(x, vars = NULL, moderator = NULL, pi = NULL,
  rawdata = FALSE, resolution = NULL, output = "plot", ...)
}
\arguments{
\item{x}{Model object.}

\item{vars}{Character vector containing the moderator names for which to plot
partial dependence plots. If empty, all moderators are plotted.}

\item{moderator}{Character. The name of a factor moderator variable in the
model object provided to \code{x}.}

\item{pi}{Numeric (0-1). What percentile interval should be plotted for the
partial dependence predictions? Defaults to NULL. To obtain a 95\% interval,
set to \code{.95}.}

\item{rawdata}{Logical, indicating whether to plot weighted raw data.
Defaults to FALSE. Uses the same weights as the model object passed to the
\code{x} argument.}

\item{resolution}{Integer vector of length two, giving the resolution of the
partial predictions. The first element indicates the resolution of the
partial predictions; for Monte-Carlo integration, the second element gives
the number of rows of the data to be sampled without replacement when
averaging over values of the other predictors.}

\item{output}{Character. What type of output should be returned? Defaults to
\code{"plot"}, which returns and plots a gtable object. To obtain a list of
\code{ggplot} objects instead, provide the argument \code{"list"}.}

\item{...}{Additional arguments to be passed to \code{marginalPrediction}.}
}
\value{
A gtable object or list of ggplot objects, see the \code{output}
parameter.
}
\description{
Plots partial dependence plots for a MetaForest model object, split by a
factor moderating variable.
}
\examples{
# Partial dependence plot for MetaForest() model:
set.seed(42)
data <- SimulateSMD(k_train = 100, model = es * x[, 1] + es * x[, 2] + es *
                                           x[, 1] * x[, 2])$training
data$X2 <- cut(data$X2, breaks = 2, labels = c("Low", "High"))
mf.random <- MetaForest(formula = yi ~ ., data = data,
                        whichweights = "random", method = "DL",
                        tau2 = 0.2450)
# Examine univariate partial dependence plot for all variables in the model:
PDP_interaction(mf.random, moderator = "X2")
}
