% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MetaForest.R
\name{print.MetaForest}
\alias{print.MetaForest}
\title{Prints MetaForest object.}
\usage{
\method{print}{MetaForest}(x, digits = NULL, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{digits}{minimal number of significant digits, see \code{print.default}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints MetaForest object.
}
\examples{
\dontshow{
set.seed(11)
data <- SimulateSMD()
mf.fixed <- MetaForest(formula = yi ~ ., data = data$training,
                      whichweights = "fixed")
mf.fixed
}
}
