% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.MetaForest.R
\name{print.summary.MetaForest}
\alias{print.summary.MetaForest}
\title{Prints summary.MetaForest object.}
\usage{
\method{print}{summary.MetaForest}(x, digits, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{digits}{minimal number of significant digits, see \code{print.default}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints summary.MetaForest object.
}
\examples{
\dontshow{
set.seed(28)
data <- SimulateSMD()
mf.random <- MetaForest(formula = yi ~ ., data = data$training,
                      whichweights = "random")
print(summary(mf.random), digits = 3)
}
}
