% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MetaForest.R
\name{summary.MetaForest}
\alias{summary.MetaForest}
\title{Summarize MetaForest object.}
\usage{
\method{summary}{MetaForest}(object, ..., digits = 2)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}

\item{digits}{the number of digits desired, defaults to 2.}
}
\description{
Summarize MetaForest object.
}
\examples{
\dontshow{
set.seed(64)
data <- SimulateSMD()
sum <- summary(mf.random <- MetaForest(formula = yi ~ .,
               data = data$training, whichweights = "random"), digits = 5)
sum
}
}
