% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preselect.R
\name{preselect}
\alias{preselect}
\title{Preselect variables for MetaForest analysis}
\usage{
preselect(formula, data, vi = "vi", study = NULL,
  whichweights = "random", num.trees = 500, mtry = NULL,
  method = "REML", tau2 = NULL, ..., replications = 100L,
  algorithm = "bootstrap")
}
\arguments{
\item{formula}{Formula. Specify a formula for the MetaForest model, for
example, \code{yi ~ .} to predict the outcome \code{yi} from all moderators
in the data. Only additive formulas are allowed (i.e., \code{x1+x2+x3}).
Interaction terms and non-linear terms are not required, as the random
forests algorithm inherently captures these associations.}

\item{data}{A data.frame containing the effect size, moderators, and the
variance of the effect size.}

\item{vi}{Character. Specify the name of the column in the \code{data} that
contains the variances of the effect sizes. This column will be removed from
the data prior to analysis. Defaults to \code{"vi"}.}

\item{study}{Character. Optionally, specify the name of the column in the
\code{data} that contains the study id. Use this when the data includes
multiple effect sizes per study. This column can be a vector of integers, or
a factor. This column will be removed from the data prior to analysis.
See \code{Details} for more information about analyzing dependent data.}

\item{whichweights}{Character. Indicate what time of weights are required.
A random-effects MetaForest is grown by specifying \code{whichweights =
"random"}. A fixed-effects MetaForest is grown by specifying
\code{whichweights = "fixed"}. An unweighted MetaForest is grown by
specifying \code{whichweights = "unif"}. Defaults to \code{"random"}.}

\item{num.trees}{Atomic integer. Specify the number of trees in the forest.
Defaults to 500.}

\item{mtry}{Atomic integer. Number of candidate moderators available for each
split. Defaults to the square root of the number moderators (rounded down).}

\item{method}{Character. Specify the method by which to estimate the residual
variance. Can be set to one of the following: "DL", "HE", "SJ", "ML", "REML",
"EB", "HS", or "GENQ". Default is "REML". See the
\code{\link[metafor:rma.uni]{metafor}} package for more information about
these estimators.}

\item{tau2}{Numeric. Specify a predetermined value for the residual
heterogeneity. Entering a value here supersedes the estimated tau2 value.
Defaults to NULL.}

\item{...}{Additional arguments are passed directly to \link[ranger]{ranger}.
It is recommended not to use additional arguments.}

\item{replications}{Integer. Number of replications to run for variable
preselection. Default: 100.}

\item{algorithm}{Character. Preselection method to apply. Currently, either
'recursive' or 'bootstrap' are available.}
}
\value{
An object of class 'mf_preselect'
}
\description{
Applies different methods to preselect variables before
conducting a \code{\link{MetaForest}} analysis.
}
\examples{
\dontrun{
data <- get(data(dat.bourassa1996))
data <- escalc(measure = "OR", ai = lh.le, bi = lh.re, ci = rh.le, di= rh.re,
               data = data, add = 1/2, to = "all")
data$mage[is.na(data$mage)] <- median(data$mage, na.rm = TRUE)
data[c(5:8)] <- lapply(data[c(5:8)], factor)
data$yi <- as.numeric(data$yi)
tmp <- preselect(formula = yi~ selection + investigator + hand_assess + eye_assess +
                        mage +sex,
          data, study = "sample",
          whichweights = "unif", num.trees = 300,
          replications = 10,
          algorithm = "recursive")
}
}
